--- src/App.css
@@ -1,66 +1,0 @@
-.mission-container {
-  display: flex;
-  background-color: lightgray;
-  width: 70%;
-  height: 100px;
-  margin: 30px;
-  margin-bottom: 0px;
-  margin-top: 0px;
-  padding: 10px;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-}
-
-.waiting-container {
-  display: flex;
-  background-color: lightgray;
-  width: 70%;
-  height: 100px;
-  margin: 80px;
-  padding: 10px;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-}
-
-.change-container {
-  height: 20px;
-  width: 130px;
-  border: none;
-  display: flex;
-  align-items: center;
-  margin-left: 66%;
-  float: right;
-}
-
-.center-container {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  height: 100vh;
-}
-
-.Header {
-  width: 100%;
-  background-color: #b0a2c8;
-  margin-top: 0;
-  margin-bottom: 50px;
-  height: 150px;
-  text-align: center;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  font-family: "Georgia", serif;
-}
-
-.changeButton {
-  border: none;
-  background-color: white;
-}
-
-.question {
-  width: 71%;
-  margin-bottom: 5px;
-}
